watchman/requirements.txt
python-telegram-bot[all]
python-dotenv